+++ editor/pages/index.tsx
